Moved reducers to their own files. Issue #6
